+++ dropwizard-core/src/main/java/com/yammer/dropwizard/config/Environment.java
@@ -5,6 +5,7 @@ import com.google.common.util.concurrent.ThreadFactoryBuilder;
 import com.sun.jersey.api.core.ResourceConfig;
 import com.sun.jersey.core.reflection.AnnotatedMethod;
 import com.sun.jersey.core.reflection.MethodList;
+import com.sun.jersey.core.spi.scanning.PackageNamesScanner;
 import com.yammer.dropwizard.AbstractService;
 import com.yammer.dropwizard.jersey.DropwizardResourceConfig;
 import com.yammer.dropwizard.jetty.JettyManaged;
@@ -15,6 +16,7 @@ import com.yammer.dropwizard.logging.Log;
 import com.yammer.dropwizard.tasks.GarbageCollectionTask;
 import com.yammer.dropwizard.tasks.Task;
 import com.yammer.metrics.core.HealthCheck;
+import org.eclipse.jetty.server.session.SessionHandler;
 import org.eclipse.jetty.servlet.FilterHolder;
 import org.eclipse.jetty.servlet.ServletHolder;
 import org.eclipse.jetty.util.component.AbstractLifeCycle;
@@ -55,6 +57,7 @@ public class Environment extends AbstractLifeCycle {
     private final ImmutableSet.Builder<EventListener> servletListeners;
     private final ImmutableSet.Builder<Task> tasks;
     private final AggregateLifeCycle lifeCycle;
+    private SessionHandler sessionHandler;
 
     /**
      * Creates a new environment.
@@ -73,6 +76,7 @@ public class Environment extends AbstractLifeCycle {
                 logHealthChecks();
                 logManagedObjects();
                 logEndpoints();
+                logTasks();
             }
         };
         this.healthChecks = ImmutableSet.builder();
@@ -106,6 +110,21 @@ public class Environment extends AbstractLifeCycle {
      */
     public void addResource(Object resource) {
         config.getSingletons().add(checkNotNull(resource));
+    }
+
+    /**
+     * Scans the packages and sub-packages of the given {@link Class} objects for resources and
+     * providers.
+     *  
+     * @param classes     the classes whose packages to scan
+     */
+    public void scanPackagesForResourcesAndProviders(Class<?>... classes) {
+        checkNotNull(classes);
+        final String[] names = new String[classes.length];
+        for(int i = 0; i < classes.length; i++) {
+            names[i] = classes[i].getPackage().getName();
+        }
+        config.init(new PackageNamesScanner(names));
     }
 
     /**
@@ -251,6 +270,10 @@ public class Environment extends AbstractLifeCycle {
      */
     public void addTask(Task task) {
         tasks.add(checkNotNull(task));
+    }
+
+    public void setSessionHandler(SessionHandler sessionHandler) {
+        this.sessionHandler = sessionHandler;
     }
 
     /**
@@ -465,11 +488,26 @@ public class Environment extends AbstractLifeCycle {
         LOG.info(stringBuilder.toString());
     }
 
+    private void logTasks() {
+        final StringBuilder stringBuilder = new StringBuilder(1024).append("\n\n");
+        
+        for (Task task : tasks.build()) {
+            stringBuilder.append(String.format("    %-7s /tasks/%s (%s)\n",
+                                               "POST", task.getName(), task.getClass().getCanonicalName()));
+        }
+        
+        LOG.info("tasks = {}", stringBuilder.toString());
+    }
+
     private MethodList annotatedMethods(Class<?> resource) {
         return new MethodList(resource, true).hasMetaAnnotation(HttpMethod.class);
     }
 
     public AbstractService<?> getService() {
         return service;
+    }
+
+    public SessionHandler getSessionHandler() {
+        return sessionHandler;
     }
 }